+++ deploy/development.yml
@@ -9,7 +9,7 @@ services:
       file: docker-compose.yml
       service: mhacks_website
     ports:
-      - "8080:8080"
+      - "3000:3000"
     volumes:
       - ..:/usr/src/app
     links:
