+++ public/pages/index.js
@@ -1,5 +1,5 @@
 import Navigator from './navigator.jsx';
-import HomePage from './home_page.jsx';
+import HomePage from './home.jsx';
 
 export {
 	Navigator,
