+++ src/app.js
@@ -35,4 +35,4 @@ app.get('/', function(req, res) {
     });
 });
 
-server.listen(process.env.PORT || 8080);
+server.listen(process.env.PORT || 3000);
